# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.1.0"
  title: Rltube-backend
# during dev, should point to your local machine
host: localhost:8080
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /{video_id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: reviews
    get:
      description: Returns the list of reviews for the video
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Response"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Add a review to this video
      parameters:
        - name: review
          in: query
          description: The name of the person to whom to say hello
          required: true
          type: string
      responses:
        "200":
          description: Success
          
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /reviews:
    # binds a127 app logic to a route
    x-swagger-router-controller: reviews
    get:
      description: Returns the list of reviews
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Response"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Add a review
      parameters:
        - name: review
          in: query
          description: The name of the person to whom to say hello
          required: true
          type: string
      responses:
        "200":
          description: Success
          
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  HelloWorldResponse:
    required:
      - msg
    properties:
      msg:
        type: string
  ErrorResponse:
    required:
      - err
    properties:
      err:
        type: number